pyproject.toml
[tool.poetry]
name = "projeto-atencao-basica-municipio"
version = "0.1.0"
description = ""
authors = ["Bruno <bruno.eng.proj@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
pandas = "^2.2.3"
sqlalchemy = "^2.0.38"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
